move image into card folder
