--- website/src/components/cards/CardTile.module.css
@@ -1,3 +1,0 @@
-.cardTile li {
-    background-color: red;
-}
